doc webdav: update copyright header, add disclaimer about python-webdav
Yes, I have been subclassing the DAV module and improving on the existing
code, regarding that we couldn't ship our own version of the whole
python-webdav library (as in v5.x).
At commit b2a2c3f95416788, on 2010-04-06, we explicitly deleted our custom
copies of the DAV module and linked to the upstream project, in fact.