[FIX] pproduct: extend default security rule to show products from ancestors companies too
This rule is not suited for most OpenERP objects, but it works
well for partners and products as these records are usually
shared between companies, and all company-specific data (such as
acounting data) is set as per-company properties. A partner or
product can thus be shared cleanly between companies.
This was already possible before by just removing the company
assigned to the product, making it globally visible.