[IMP] share,portal: improve sharing with portal group
Use a new strategy: documents are shared with a
temporary group, and when everything is setup,
the access rights are copied to the groups we are
sharing with, and the rules are transferred to them.
The temporary group is then deleted.
This works well for multiple shares, are both
access rights and rules are additive in this 
fashion.