[FIX] base_action_rule: dumb patch to avoid repeated calls to related cron task
base.action.rule uses a trick to piggyback on the
first execution of ir.cron's poolJobs() to register
its ORM hooks. It was currently doing that every
single time poolJobs() was called, leading to
incessant querying the ORM objects for which a
base action rule exist. 
Now it will do it only once, and then respect the
scheduled executions.
Soon this mechanism will be removed and replaced
by a proper event handler for some 'ORM finished init'
event, to be fired by the server.