[FIX] contracts: use correct quantity field to avoid mismatching values
`quantity` is a recursive sum of quantities of all types, including child accounts,
while `hours_quantity` is the local sum of quantities in `general` analytic journals.
The latter is what corresponds to the `Units consumed` value for contracts, and what
is used to compute the `remaining_hours`, so we must used that.