[FIX] portal_anonymous: `token` and `login` params are in the URL fragment, not in the query string
The previous test was done on the query string, in which
`token` and `login` are never found, preventin users
from seeing their reset password form.
The test was also incorrectly checking for the presence
of both a `token` and `login`, while any of those should
disable the anonymous mode (sometimes only `token` is
present, e.g. for password reset).