[FIX] base_report_designer: fix upload code in plugin to avoid broken reports
It is the call to upload_report() that triggers the registration
of the new reports in the system, as report services.
Unfortunately the `header` property of the report is cached in
the report service and taken from its value at
registration time. So that value *must* be written before
calling upload_report().

Also force the `Corporate Header` to be checked by default
as this is what users want in most cases, and forgetting
it at report creation makes it very hard to set afterwards,
as it is cached in the service.

Updated plugin binary as well.