[FIX] pricelist: correctly take into account uom when computing pricelists based on supplier price on product form (opw 595881)
The previous behaviour used the uom of product while it could be a different one selected (by default the purchase unit of measure for purchase orders).
This was an issue especially when having different uom with supplier info lines setting degressive prices. The price should be computed based on selected uom and not the product uom.